Directly stream backup to the Ansible host
Instead of creating the backup on the galera container and then rsyncing
it to the deployment host, directly stream it to the Ansible host. This
makes the workflow a lot simpler and allows using a restricted SSH
authentication key which is only allowed to execute the backup command
from the Ansible host.