Optionally setup SSH key authentication
In recent versions of OSA it's no longer necessary to have the OSA SSH
key installed on all containers as it uses a special connection plugin
to connect to the container through the LXC host.

If the OSA SSH key is not present on the containers for unrestricted
access this allows configuration of the key to only execute the backup
command from the Ansible host.